Move all menu-type widgets to their dedicated folder
