+++ applications/main/busy/scenes/busy_scene_setup.c
@@ -1,6 +1,6 @@
 #include "../busy_i.h"
 
-#include <gui/modules/menu.h>
+#include <gui/modules/menus/menu.h>
 #include <gui/modules/anim_player.h>
 
 #define ITEM_LABEL_TIMER      "Timer"
