+++ applications/main/busy/scenes/busy_scene_setup_timer.c
@@ -2,7 +2,7 @@
 
 #include <furi_hal_nvm.h>
 
-#include <gui/modules/var_item_list.h>
+#include <gui/modules/menus/var_item_list.h>
 
 typedef enum {
     VarItemListIdMode,
