+++ applications/services/gui/modules/menus/base/menu_base_i.h
@@ -1,5 +1,5 @@
 #include "menu_base.h"
-#include "../widget_i.h"
+#include "../../../widget_i.h"
 
 struct MenuBase {
     Widget base;
