+++ applications/services/gui/modules/menus/menu.c
@@ -1,6 +1,7 @@
 #include "menu.h"
 
-#include "menu_base_i.h"
+#include "base/menu_base_i.h"
+
 #include <gui/modules/anim_player_i.h>
 
 #include <lvgl/src/core/lv_obj_class_private.h>
