+++ applications/services/gui/modules/menus/submenu.c
@@ -1,5 +1,6 @@
 #include "submenu.h"
-#include "menu_base_i.h"
+
+#include "base/menu_base_i.h"
 
 #include <gui/modules/label.h>
