+++ applications/services/gui/modules/menus/var_item_list.c
@@ -1,6 +1,6 @@
 #include "var_item_list.h"
 
-#include "menu_base_i.h"
+#include "base/menu_base_i.h"
 
 #include <lvgl/src/core/lv_obj_class_private.h>
 #include <lvgl/src/widgets/label/lv_label_private.h>
