+++ applications/settings/about/scenes/about_scene_libs_list.c
@@ -1,6 +1,6 @@
 #include "../about.h"
 #include <settings_helpers/gui_params.h>
-#include <gui/modules/submenu.h>
+#include <gui/modules/menus/submenu.h>
 
 typedef enum {
     SceneEventLibSelectStart = AppEventSceneEventsStart,
