+++ applications/settings/ble_settings/scenes/scene_forget_device.c
@@ -1,8 +1,8 @@
 #include "../ble_settings.h"
 #include <settings_helpers/gui_params.h>
 
-#include <gui/modules/var_item_list.h>
-#include <gui/modules/submenu.h>
+#include <gui/modules/menus/var_item_list.h>
+#include <gui/modules/menus/submenu.h>
 
 typedef enum {
     SceneEventRemovePairing = AppEventSceneEventsStart,
