+++ applications/settings/system_settings/scenes/system_settings_scene_debug.c
@@ -1,6 +1,6 @@
 #include "../system_settings.h"
 
-#include <gui/modules/var_item_list.h>
+#include <gui/modules/menus/var_item_list.h>
 #include <furi_hal_nvm.h>
 #include <sysctl/sysctl.h>
