+++ applications/settings/system_settings/scenes/system_settings_scene_power_menu.c
@@ -1,7 +1,7 @@
 
 #include "../system_settings.h"
 
-#include <gui/modules/submenu.h>
+#include <gui/modules/menus/submenu.h>
 
 typedef enum {
     SceneEventShutDown = AppEventSceneEventsStart,
