+++ applications/settings/time_settings/scenes/scene_menu.c
@@ -2,7 +2,7 @@
 #include <settings_helpers/gui_params.h>
 #include <furi_hal_rtc.h>
 
-#include <gui/modules/submenu.h>
+#include <gui/modules/menus/submenu.h>
 
 #define TAG "TIME_MENU"
